Fix: 베이스볼 게임 버그 수정 완료
