+++ src/main/java/baseball/BaseballMain.java
@@ -4,6 +4,6 @@ public class BaseballMain {
     public static void main(String[] args) {
         BaseballGame baseballGame = new BaseballGame();
         baseballGame.startGame();
-        baseballGame.endGame();
+        baseballGame.reGame();
     }
 }
